Clarify the end of Step 6
It took me a few minutes to figure out that the authentication was done, because I thought the page would automatically load the screen that you show next in the instructions, which isn't the case. I was searching and searching for the settings button (three horizontal lines) on the Profile Settings page, but it doesn't exist there. The user needs to exit the Profile Settings page to get back to the main screen to access the Nightscout settings button (three horizontal lines).