Update update_user.md
On ENABLE line, entered "bage" as replacement for duplicate word "basal", because (1) ENABLE note says "avoid duplicates", (2) "bage" has that position in the line of words on the "new_user.md" page, that is, following "sage" and "cage". Plus a minor change in ENABLE note to more accurately reference the code "above" rather than "below".