describe facets of pre-meal
* reword to be little less redundant and avoid parentheticals
* added two sections to facilitate possible further additions to these facets

This patch is intended to help with a discussion that started on Facebook
regarding admonishments on the pre-meal feature.  It's not possible to unset
however, there were some concerns regarding assessing the impact with
implications for how to choose correction values.  This change mentions a way
to limit the impact of pre-meal while trying to avoid raising additional
concerns.  In addition, we connect to the intent and origin of pre-meal with an
admonition warning against low values.