Clarify build step for Loop-dev
When I followed these instructions yesterday for Loop-dev, I nearly started the Build Loop page over from scratch, instead of skipping to the "Build Loop" section of the Build Loop page. Tried to make it clearer that you are actually skipping directly to a section of that page if you're building Loop-dev.