Change explanation of insulin activity
First part of the proposed change:

Original text of first sentence cannot be right - delta BG is the velocity or rate of change in blood glucose due to insulin, so this cannot also be the definition of insulin activity. Have proposed a change that inverts this part of the sentence, which I think is consistent with the relationship of IA and Delta BG in the formula; but there may be a better physiological explanation.

Second part of the proposed change:

My change in the second sentence I'm not fully comfortable with, but the original sentence I found confusing. What is the EGP? How does it account for the EGP? Does 'basal' in this sentence relate to scheduled basal or variations from it due to temporary basals?

Initially I came here just to propose an edit so that EGP was defined (unlike ISF, Loop etc it doesn't get the 'definition on hover over' treatment) but on closer reading the sentence didn't seem right