fix(features): Fix broken equations in open-loop feature
As `MTB` is a glossary item, it needs to be “broken down”
to prevent the preprocessor from finding it,
and inserting the tooltip HTML code which breaks the equation.

See https://github.com/LoopKit/loopdocs/pull/892
I used this bulky workaround to explode the glossary acronym into multiple parts
(at least 2) to hide it from the glossary acronym preprocessor.
Replace `MTB` with `\mathit{MT}\mathit{B}`.

ℹ️ Where `\mathit{MT}` italicizes `MT`.